graph implementation. current progress commit
